Create AutoJoin.mrc
Initial release of the AutoJoin script for mIRC. Based on my AutoJoin module for ZNC.

-=-=-=-=-=-=-=-=-=-= Script Readme =-=-=-=-=-=-=-=-=-=-

AutoJoin Script for Undernet IRC Network
By QueenElsa                Version: 1.9

USAGE: Load this script into a blank new remotes file on your client. Script uses X's autoinvites for its primary function. There are no options to configure within this script. Everything is done through X. it may be expanded to include other bots and services, as this script is not complicated. 

STAFF NOTE: If you're an Undernet Staff Member, a special version of this script exists and may be available to you. Contact QueenElsa for more information.